requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
passlib
python-jose
torch
torchvision
Pillow
